+++ app/src/main/java/eu/kanade/tachiyomi/ui/player/controls/PlayerControls.kt
@@ -90,6 +90,7 @@ import eu.kanade.tachiyomi.ui.player.settings.AudioPreferences
 import eu.kanade.tachiyomi.ui.player.settings.DecoderPreferences
 import eu.kanade.tachiyomi.ui.player.settings.GesturePreferences
 import eu.kanade.tachiyomi.ui.player.settings.PlayerPreferences
+import eu.kanade.tachiyomi.ui.player.settings.SubtitleAssOverride
 import eu.kanade.tachiyomi.ui.player.settings.SubtitleJustification
 import eu.kanade.tachiyomi.ui.player.settings.SubtitlePreferences
 import kotlinx.collections.immutable.persistentListOf
@@ -717,7 +718,7 @@ fun PlayerControls(
         val subColor by viewModel.mpv.propFlow<String>("sub-color").collectAsState()
         val subBorderColor by viewModel.mpv.propFlow<String>("sub-outline-color").collectAsState()
         val subBackgroundColor by viewModel.mpv.propFlow<String>("sub-back-color").collectAsState()
-        val overrideAssSubs by viewModel.mpv.propFlow<Boolean>("sub-ass-override").collectAsState()
+        val overrideAssSubs by viewModel.mpv.propFlow<String>("sub-ass-override").collectAsState()
         val subScale by viewModel.mpv.propFlow<Float>("sub-scale").collectAsState()
         val subPos by viewModel.mpv.propFlow<Int>("sub-pos").collectAsState()
         val deband by decoderPreferences.debanding().collectAsState()
@@ -752,7 +753,8 @@ fun PlayerControls(
                 SubColorType.Background -> subBackgroundColor?.toColorInt()
                     ?: subtitlePreferences.backgroundColorSubtitles().get()
             },
-            overrideAssSubs = overrideAssSubs ?: subtitlePreferences.overrideSubsASS().get(),
+            overrideAssSubs = overrideAssSubs?.let { SubtitleAssOverride.byValue(it) }
+                ?: subtitlePreferences.overrideSubsASS().get(),
             subScale = subScale ?: subtitlePreferences.subtitleFontScale().get(),
             subPos = subPos ?: subtitlePreferences.subtitlePos().get(),
             onSubBoldChange = {
@@ -806,7 +808,7 @@ fun PlayerControls(
                 }
             },
             onOverrideAssSubsChange = {
-                viewModel.mpv.setPropertyBoolean("sub-ass-override", it)
+                viewModel.mpv.setPropertyString("sub-ass-override", it.value)
                 subtitlePreferences.overrideSubsASS().set(it)
             },
             onSubScaleChange = {
@@ -831,8 +833,9 @@ fun PlayerControls(
                 subtitlePreferences.subtitleFontScale().deleteAndGet().let {
                     viewModel.mpv.setPropertyFloat("sub-scale", it)
                 }
-                subtitlePreferences.overrideSubsASS().delete()
-                viewModel.mpv.setPropertyString("sub-ass-override", "scale")
+                subtitlePreferences.overrideSubsASS().deleteAndGet().let {
+                    viewModel.mpv.setPropertyString("sub-ass-override", it.value)
+                }
             },
             subDelayMsPrimary = subDelay?.times(1000)?.roundToInt() ?: subDelayPref,
             subDelayMsSecondary = subDelaySecondary?.times(1000)?.roundToInt() ?: subDelaySecondaryPref,