+++ app/src/main/java/mihon/core/migration/migrations/Migrations.kt
@@ -45,4 +45,5 @@ val migrations: List<Migration>
         PrefLangMigration(),
         RenameKeysMigration(),
         MoveCoverCacheMigration(),
+        SubtitleAssEnumMigration(),
     )